Adding initial files to the repository
main.py -- source for Python app that makes suggestions for the next word to guess while playing Wordle. It is still the player's responsibility to make a good choice, but they should never come up blank.

Scrabble Words.txt -- a text file containing a large list of English words.

Setup.txt -- a text file with a few setting for controlling the operation of the application.

WORDLEs.txt -- this file holds a history of WORDLE answers and user guesses, which are used to give the user better guidance in the current game (e.g. separating unused words from previously-used words that are unlikely to be used again).