Bug fix in Wordle app -- failed to find words in ROATE ELFIN NEWEL KNELL sequence
When the word was KNELL, roate/elfin/newel reported that there were no words left.
It was failing when first "e" was reported to be in the right_wrong place, while second "e" was wrong_wrong. Changed it to look at the count of the wrong letters to make sure it was zero before clearing this letter from other positions.